Revert "Move fixtures to conftest.py"
This reverts commit 3d02b15d
